Works again in PyCharm thanks to smarter use of inspect
Rather than parsing the code context for the invocation of
the save() function, we now match up arguments to variables
in the caller's namespace by their id(). This means that save()
now works even when code context is not available, such as
may be the case in some interactive REPLs, including, notably
at least one version of PyCharm.